Update main.swift
Minor tweak to change falconctl to falcond for the Crowdstrike check. The falcond is the the running daemon on the system, whereas falconctl is more of a utility binary used when configuring new agents or retrieving stats. I also tweaked the last check for falcon to include the isDirectory instance method.